% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_style.R
\name{align_lc}
\alias{align_lc}
\title{Left-center align}
\usage{
align_lc(object)
}
\arguments{
\item{object}{a \code{flextable} object.}
}
\value{
a \code{flextable} object with left-centered columns
}
\description{
This function takes care of some tedious stylistic
changes that users may want to automate. Specifically,
left aligning the far left column and center aligning
every other column.
}
\examples{

library(flextable)
library(dplyr)

iris \%>\%
  group_by(Species) \%>\%
  slice(1:3) \%>\%
  as_grouped_ft() \%>\%
  theme_vanilla() \%>\%
  align_lc() \%>\%
  pad_groups()

}
