% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_style.R
\name{as_grouped_ft}
\alias{as_grouped_ft}
\title{Flextables}
\usage{
as_grouped_ft(data, hide_group_label = TRUE)
}
\arguments{
\item{data}{a grouped dataset.}

\item{hide_group_label}{a logical value. if \code{TRUE}, group label
will not be rendered, only level/value will be rendered.}
}
\value{
a \code{flextable} object with an additional \code{group_names}
object attached.
}
\description{
Simple workflows involving \code{\link[flextable:flextable]{flextable::flextable()}}.
In particular, grouped data from the \code{\link[dplyr:group_by]{dplyr::group_by()}}
package are handled so that group columns are automatically
turned into groups of rows in the table.
}
\examples{

library(flextable)
library(dplyr)

iris \%>\%
  group_by(Species) \%>\%
  slice(1:3) \%>\%
  as_grouped_ft() \%>\%
  theme_vanilla() \%>\%
  align_lc() \%>\%
  pad_groups()

}
