% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_style.R
\name{flip_names}
\alias{flip_names}
\title{Flip names}
\usage{
flip_names(df, pattern)
}
\arguments{
\item{df}{a data frame}

\item{pattern}{a string indicating the pattern that separates
column header patterns}
}
\value{
a data frame with names flipped around \code{pattern}
}
\description{
This function is a convenient re-name tool in case
you want to use \link{header_split} but the names of your variables
are set up in the opposite order of what \link{header_split} expects.
}
\examples{

df = data.frame(a_c = 1, a_b = 2)

flip_names(df, '_')


}
