% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_style.R
\name{ft}
\alias{ft}
\title{Flextable}
\usage{
ft(
  data,
  ...,
  col_keys = names(data),
  cwidth = 0.75,
  cheight = 0.25,
  defaults = list(),
  use_col_labels = TRUE,
  theme_fun = flextable::theme_booktabs
)
}
\arguments{
\item{data}{dataset}

\item{...}{name-value pairs for column names and column labels}

\item{col_keys}{columns names/keys to display. If some column names
are not in the dataset, they will be added as blank columns.}

\item{cwidth}{initial width to use for cell sizes in inches.}

\item{cheight}{initial height to use for cell sizes in inches.}

\item{defaults}{a list of default values for formats, supported options
are fontname, font.size, color and padding.}

\item{use_col_labels}{a logical value. If \code{TRUE}, existing \code{label}
attributes on columns in \code{data} will be applied to their corresponding
table columns.}

\item{theme_fun}{a function theme to apply before returning the
flextable. set to NULL for none.}
}
\value{
a \code{flextable} object
}
\description{
this is a basic wrapper of \code{\link[flextable:flextable]{flextable::flextable()}} that
gives some additional functionality for data with labelled columns.
The inclusion of a \code{...} argument allows you to include name-value
pairs for column names and column labels. This function will also
automatically pick up on column labels in \code{tblString} objects.
}
\examples{

# default flextable behavior
ft(iris[1:5, ])
# adding labels in flextable call
ft(iris[1:5, ], 'Sepal Width' = Sepal.Width, cwidth = 1.5)

}
