% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_style.R
\name{header_split}
\alias{header_split}
\title{Flextable headers}
\usage{
header_split(object, pattern = "_", ...)
}
\arguments{
\item{object}{a \code{flextable} object}

\item{pattern}{a string indicating the pattern that separates
column header patterns in your \code{flextable}'s \code{col_keys},}

\item{...}{named inputs of the form ||table header label|| =
||column name chunk||. Examples given below should make this
a little more clear.}
}
\value{
a \code{flextable} object
}
\description{
This function helps to simplify table headers, assuming
the table's \code{col_keys} have the following structure:
}
\details{
||LOWER COLUMN NAME|| ||PATTERN|| ||HIGHER COLUMN NAME|| ||PATTERN|| ...

for example, 'sepal_width' will be split into sepal and width if
the pattern is '_' and then a lower header row will feature the
text 'sepal', and an upper header row will show 'width'.

But what are \code{col_keys} of a \code{flextable}? By default, they are
names of the \code{data} used to make the \code{flextable}. For a grouped
\code{flextable}, group names are not included in \code{col_keys}.

This function also lets you enter mappings from the column names
in your data to the desired column headers you'd like to see in
the table.
}
\examples{

df <- data.frame(a_c = c("easy", "make"), b_c = c("to", "now!"))
ft <- flextable::flextable(df)
ft <- header_split(ft, pattern = '_', c = 'Headers', a = 'are', b = 'so')
ft <- flextable::theme_box(ft)
ft <- flextable::align(ft, align = 'center', part = 'all')
ft


}
