% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_style.R
\name{pad_groups}
\alias{pad_groups}
\title{Flextables}
\usage{
pad_groups(
  object,
  grp_color = "grey80",
  pad_width = 10,
  bold = FALSE,
  italic = TRUE
)
}
\arguments{
\item{object}{a \code{flextable} object with group names. To create this
type of object, use \link{ft_grouped_by}.}

\item{grp_color}{a character value (or vector of length equal to the
number of groups in \code{data}) indicating the background color that
will be applied to rows that label groups.}

\item{pad_width}{a numeric value indicating the amount of space added
to child rows in each group.}

\item{bold}{a logical value. If \code{TRUE}, then the group labels
will be written in bold text.}

\item{italic}{a logical value. If \code{TRUE}, then the group labels
will be written in italic text.}
}
\value{
a \code{flextable} object with grouped rows padded.
}
\description{
the \code{ft} functions help accelerate workflows involving
\code{\link[flextable:flextable]{flextable::flextable()}}. In particular, grouped data
from the \code{\link[dplyr:group_by]{dplyr::group_by()}} package are handled so that
group columns are automatically turned into groups of rows
in the table.
}
\examples{

library(flextable)
library(dplyr)

iris \%>\%
  group_by(Species) \%>\%
  slice(1:3) \%>\%
  ft_grouped_by() \%>\%
  theme_vanilla() \%>\%
  align_lc() \%>\%
  pad_groups()

}
