% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smry_funs.R
\name{smry_meanSD}
\alias{smry_meanSD}
\alias{smry_medianIQR}
\alias{smry_countP}
\alias{smry_irateCI}
\title{Named summaries}
\usage{
smry_meanSD(x, name = "Mean (SD)", na.rm = TRUE)

smry_medianIQR(x, name = "Median [IQR]", na.rm = TRUE, lwr = 0.25, upr = 0.75)

smry_countP(x, use_na = TRUE, include_first_cat = TRUE, na_level = "Missing")

smry_irateCI(
  time,
  status,
  name = "Incidence rate (95\% CI)",
  unit = 1000,
  level = 0.95
)
}
\arguments{
\item{x}{a numeric vector}

\item{name}{a character value indicating the name of the output.}

\item{na.rm}{a logical value indicating whether NA values
should be stripped before the computation proceeds.}

\item{lwr}{a numeric value indicating the lower quantile to report.}

\item{upr}{a numeric value indicating the upper quantile to report.}

\item{use_na}{a logical value indicating whether another category
should be included showing the count and percent of missing values.}

\item{include_first_cat}{a logical value indicating whether you'd
like to include the first category (i.e., the reference group)
when you present counts and percents for a categorical variable.}

\item{na_level}{a character value indicating how missing categories
should be labeled. This is only relevant if \code{use_na = TRUE}}

\item{time}{a numeric vector of time to event values}

\item{status}{a numeric vector of 0s and 1s. Values of 0 should indicate
censoring, and values of 1 indicate observed events.}

\item{unit}{a numeric value indicating the unit of incidence. For example,
\code{unit = 1000} will compute incidence rate per 1,000 person-years.}

\item{level}{the level of conidence limits. For example, \code{level = 0.95}
leads to 95\% confidence limits.}
}
\value{
a named character value
}
\description{
Tabulation of data into character values that
are easily passed into tables. These functions are structured
so that users can easily create their own. A named summary
function will do the following:
\itemize{
\item take a numeric vector as input
\item return a \strong{named} character vector as an output
}

The name of the output is intended to be a label for
the output in your tables. When these functions are used
within the \code{dabulate} function, these qualifications work
with the structure of \code{dplyr} and \code{purrr}.
}
\examples{
smry_meanSD(rnorm(100))
smry_medianIQR(rnorm(100))


}
