% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblHelpers-package.R
\docType{package}
\name{tblHelpers-package}
\alias{tblHelpers}
\alias{tblHelpers-package}
\title{tblHelpers: Tabulation Helper Functions}
\description{
fabulate provides a set of functions to help accelerate
  workflows that tabulate data.
}
\author{
\strong{Maintainer}: Byron Jaeger \email{bcjaeger@uab.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
