% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_style.R
\name{width_lc}
\alias{width_lc}
\title{Left-center widths}
\usage{
width_lc(object, width_l = 2, width_c = 1.25)
}
\arguments{
\item{object}{a \code{flextable} object.}

\item{width_l}{a numeric value indicating the width of the far left
column in the table (i.e., what normally contains labels)}

\item{width_c}{a numeric value indicating the width of every
column other than the far left column.}
}
\value{
a \code{flextable} object with adjusted widths
}
\description{
Left-center widths
}
\examples{

library(flextable)
library(dplyr)

iris \%>\%
  group_by(Species) \%>\%
  slice(1:3) \%>\%
  as_grouped_ft() \%>\%
  theme_vanilla() \%>\%
  align_lc() \%>\%
  pad_groups()

}
