% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_style.R
\name{header_split}
\alias{header_split}
\title{Flextable headers}
\usage{
header_split(object, ..., pattern = "_", front_to_back = TRUE)
}
\arguments{
\item{object}{a \code{flextable} object}

\item{...}{named inputs of the form \verb{<table header label>} =
\verb{<column name chunk>}. Examples given below should make this
a little more clear.}

\item{pattern}{a string indicating the pattern that separates
column header patterns in your \code{flextable}'s \code{col_keys}.}

\item{front_to_back}{a logical value. If \code{TRUE}, then the table headers
are constructed from front to back, meaning variable names \code{a_b} and
\code{a_c} will be turned into a header value \code{a} comprising lower rows
\code{b} and \code{c}. If \code{FALSE}, then the names \code{b_a} and \code{c_a} would be needed
to make the same structure.}
}
\value{
a \code{flextable} object
}
\description{
This function helps to simplify table headers, assuming
the table's \code{col_keys} have the following structure:
}
\details{
||LOWER COLUMN NAME|| ||PATTERN|| ||HIGHER COLUMN NAME|| ||PATTERN|| ...

for example, 'sepal_width' will be split into sepal and width if
the pattern is '_' and then a lower header row will feature the
text 'sepal', and an upper header row will show 'width'.

But what are \code{col_keys} of a \code{flextable}? By default, they are
names of the \code{data} used to make the \code{flextable}. For a grouped
\code{flextable}, group names are not included in \code{col_keys}.

This function also lets you enter mappings from the column names
in your data to the desired column headers you'd like to see in
the table.
}
\examples{

df <- data.frame(a_c = c("made", "column"), b_c = c("with", "names"))
ft <- flextable::flextable(df)
ft <- header_split(ft, pattern = '_',
  c = 'Headers', a = 'are', b = 'all')
ft <- flextable::theme_box(ft)
ft <- flextable::align(ft, align = 'center', part = 'all')
ft

}
